[KFI]test(pickers-react): moved mock to it's own folder
